+++ tests/ValdiTests/RulesBuilderTest.php
@@ -28,11 +28,11 @@ class RulesBuilderTest extends PHPUnit_Framework_TestCase
     public function testRulesBuilding()
     {
         $read = RulesBuilder::create()
-            ->addFieldRule('a', 'required')
-            ->addFieldRule('b', 'inThePast')
-            ->addFieldRule('a', 'min', 42)
-            ->addFieldRule('c', 'slug')
-            ->addFieldRule('b', 'between', 5, 17)
+            ->field('a', 'required')
+            ->field('b', 'inThePast')
+            ->field('a', 'min', 42)
+            ->field('c', 'slug')
+            ->field('b', 'between', 5, 17)
             ->build();
         $expected = [
             'a' => [['required'], ['min', 42]],
@@ -45,8 +45,8 @@ class RulesBuilderTest extends PHPUnit_Framework_TestCase
     public function testCreatedRules()
     {
         $rules = RulesBuilder::create()
-            ->addFieldRule('a', 'required')
-            ->addFieldRule('b', 'min', 5)
+            ->field('a', 'required')
+            ->field('b', 'min', 5)
             ->build();
         $data = ['a' => 'abc', 'b' => 6];
         $validator = new Validator();
@@ -62,9 +62,9 @@ class RulesBuilderTest extends PHPUnit_Framework_TestCase
     {
         $validator = new Validator();
         $builder = RulesBuilder::create();
-        $elementRules = $builder->addRule('min', 5)->build();
+        $elementRules = $builder->rule('min', 5)->build();
         $rules = $builder
-            ->addFieldRule('a', 'collection', $validator, $elementRules)
+            ->field('a', 'collection', $validator, $elementRules)
             ->build();
         $data = ['a' => [6, 7, 8]];
         $read = $validator->isValid($rules, $data);